newsguard-hackathon/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
scikit-learn==1.5.1
pandas==2.2.2
numpy==2.0.1
nltk==3.9.1
